+++ ServerBase/server_base.h
@@ -29,7 +29,7 @@ public:
 	ContextPool* GetContextPool();
 
 	// get peer server by id or name
-	PEER_SERVER GetPeerServer(uint16 iServerId);
+	PEER_SERVER GetPeerServer(uint8 iServerId);
 	PEER_SERVER GetPeerServer(const TCHAR* strServerName);
 
 	// time control
